backend/srnet/pyproject.toml
[project]
name = "srnet"
version = "0.1.0"
description = "Scene Text Replacement Network API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "python-dotenv==1.0.0",
    "matplotlib==3.7.3",
    "numpy==1.24.4",
    "opencv-python==4.5.4.60",
    "pandas==2.0.3",
    "pillow==10.0.1",
    "protobuf==4.24.3",
    "regex==2023.10.3",
    "requests==2.28.2",
    "scikit-image==0.21.0",
    "scipy==1.10.1",
    "sentencepiece==0.1.99",
    "setuptools==60.2.0",
    "torch==1.13.1",
    "torchvision==0.14.1",
    "tqdm==4.65.2",
    "transformers==4.33.3",
]